Document bnc#798987
Document that when using virt-viewer on Dom0 to install the required
packages before. This issue might be only relevant for SLES 10 SP4 since
our SLES 11 SP2 XEN Host Mode appliance doesn't seem to be shipped with
virt-viewer by default but if people want it there they also have to
install it so I think it is fine that way.