remove VERSION from master
Builds from master will contain a timestamp. This will make it easier
to build it on a continuous way, that is, every commit will create
a new package. If we were using a VERSION file instead of a timestamp,
we will end up with lots of versions and lots of tags, what will have
made git very messy. Moreover, we would have to maintain the VERSION
file and that could easilly create conflicts in the continuous build
system and in git tags.