use bs-update to update the package into the build service
bs-update (https://github.com/roman-neuhauser/bs-update) is a Build
Service Update script. Its purpose is to update the tarball and the
spec file in the OpenSUSE Build Service (http://build.opensuse.org).

If we use bs-update, we do not need to have a "make_snapshot" script
that creates the snapshot and updates the spec file, since this will
be done by bs-update.

Because of that, the changes are:

- For bs-update to work, the version in the spec file should be
__VERSION__ instead of #VERSION#

- Remove the VERSION file in master since we don't use the makesnapshot
script anymore.

- use gzip instead of bzip2 because that is what bs-update productes